Move all functions to two seperate files.
