--- src/getStorage.js
@@ -1,6 +1,0 @@
-const getStorage = () => {
-  const storage = JSON.parse(localStorage.getItem('storage'));
-  return (storage !== null) ? storage : [];
-};
-
-export { getStorage as default };
