--- src/orderTasks.js
@@ -1,7 +1,0 @@
-const orderedTasks = (tasks) => {
-  for (let i = 0; i < tasks.length; i += 1) {
-    tasks[i].index = i + 1;
-  }
-};
-
-export { orderedTasks as default };
