--- src/removeCompletedTasks.js
@@ -1,13 +1,0 @@
-import updateStorage from './updateStorage.js';
-import showList from './showList.js';
-import orderTasks from './orderTasks.js';
-
-const removeCompletedTasks = (tasks) => {
-  tasks = tasks.filter((task) => task.completed === false);
-  orderTasks(tasks);
-  updateStorage(tasks);
-  showList(tasks);
-  return tasks;
-};
-
-export { removeCompletedTasks as default };
