--- src/showList.js
@@ -1,42 +1,0 @@
-const showList = (tasksList) => {
-  const listDiv = document.getElementById('list');
-  listDiv.innerHTML = '';
-  for (let index = 0; index < tasksList.length; index += 1) {
-    const task = document.createElement('li');
-    const remove = document.createElement('i');
-    const edit = document.createElement('i');
-    const descriptionContainer = document.createElement('div');
-    const description = document.createElement('p');
-    const editDescription = document.createElement('input');
-    if (tasksList[index].completed) {
-      const checked = document.createElement('i');
-      checked.className = 'fas fa-check';
-      descriptionContainer.appendChild(checked);
-      description.classList.add('completed');
-    } else {
-      const check = document.createElement('input');
-      check.type = 'checkbox';
-      check.className = 'check';
-      descriptionContainer.appendChild(check);
-    }
-    edit.className = 'fas fa-edit d-none';
-    editDescription.type = 'text';
-    editDescription.id = 'edit-Description';
-    editDescription.className = 'edit-description d-none';
-    task.id = tasksList[index].index;
-    remove.className = 'fas fa-trash';
-    descriptionContainer.classList.add('description-container');
-    description.className += ' description';
-    description.innerHTML += ` ${tasksList[index].description}`;
-    descriptionContainer.appendChild(description);
-    descriptionContainer.appendChild(editDescription);
-    task.className = 'task';
-    task.setAttribute('draggable', 'true');
-    task.appendChild(descriptionContainer);
-    task.appendChild(remove);
-    task.appendChild(edit);
-    listDiv.appendChild(task);
-  }
-};
-
-export { showList as default };
