--- src/updateStatus.js
@@ -1,7 +1,0 @@
-const updateStatus = (list, id, completed) => {
-  list.forEach((task) => {
-    if (task.index === id) task.completed = completed;
-  });
-};
-
-export { updateStatus as default };
