--- src/updateStorage.js
@@ -1,5 +1,0 @@
-const updateStorage = (storage) => {
-  localStorage.setItem('storage', JSON.stringify(storage));
-};
-
-export { updateStorage as default };
